More cleaning and comments
