+++ nanoFramework.Graphics/System/WindowCollection.cs
@@ -11,8 +11,6 @@ using System.Diagnostics;
 
 namespace nanoFramework.UI
 {
-    #region WindowCollection class
-
     /// <summary>
     /// WindowCollection can be used to interate over all the windows that have been
     /// opened in the current application.
@@ -20,14 +18,12 @@ namespace nanoFramework.UI
     //CONSIDER: Should this be a sealed class?
     public sealed class WindowCollection : ICollection
     {
-        //------------------------------------------------------
-        //
-        //   Public Methods
-        //
-        //------------------------------------------------------
+        private ArrayList _list;
+
         #region Public Methods
+
         /// <summary>
-        /// Default Constructor
+        /// Default constructor for the WindowCollection class.
         /// </summary>
         public WindowCollection()
         {
@@ -35,6 +31,10 @@ namespace nanoFramework.UI
             _list.Capacity = 1;
         }
 
+        /// <summary>
+        /// Internal constructor for the WindowCollection class that takes a count parameter.
+        /// </summary>
+        /// <param name="count">The count of windows to initialize the collection with.</param>
         internal WindowCollection(int count)
         {
             //Debug.Assert(count >= 0, "count must not be less than zero");
@@ -44,15 +44,13 @@ namespace nanoFramework.UI
 
         #endregion Public Methods
 
-        //------------------------------------------------------
-        //
-        //   Operator overload
-        //
-        //------------------------------------------------------
         #region Operator overload
+
         /// <summary>
-        /// Overloaded [] operator to access the WindowCollection list
+        /// Gets the Window object at the specified index.
         /// </summary>
+        /// <param name="index">The zero-based index of the window to retrieve.</param>
+        /// <returns>The Window object at the specified index.</returns>
         public Window this[int index]
         {
             get
@@ -63,16 +61,12 @@ namespace nanoFramework.UI
 
         #endregion Operator overload
 
-        //------------------------------------------------------
-        //
-        //   IEnumerable implementation
-        //
-        //------------------------------------------------------
         #region IEnumerable implementation
+
         /// <summary>
-        /// GetEnumerator
+        /// Returns an enumerator that iterates through the WindowCollection.
         /// </summary>
-        /// <returns></returns>
+        /// <returns>An IEnumerator object that can be used to iterate through the collection.</returns>
         public IEnumerator GetEnumerator()
         {
             return _list.GetEnumerator();
@@ -80,34 +74,30 @@ namespace nanoFramework.UI
 
         #endregion IEnumerable implementation
 
-        //--------------------------------------------------------
-        //
-        //   ICollection implementation (derives from IEnumerable)
-        //
-        //--------------------------------------------------------
         #region ICollection implementation
+
         /// <summary>
-        /// CopyTo
+        /// Copies the entire WindowCollection to a compatible one-dimensional Array, starting at the specified index of the target array.
         /// </summary>
-        /// <param name="array"></param>
-        /// <param name="index"></param>
+        /// <param name="array">The one-dimensional Array that is the destination of the elements copied from WindowCollection. The Array must have zero-based indexing.</param>
+        /// <param name="index">The zero-based index in the array at which copying begins.</param>
         void ICollection.CopyTo(Array array, int index)
         {
             _list.CopyTo(array, index);
         }
 
         /// <summary>
-        /// CopyTo
+        /// Copies the elements of the WindowCollection to an Array, starting at a particular Array index.
         /// </summary>
-        /// <param name="array"></param>
-        /// <param name="index"></param>
+        /// <param name="array">The one-dimensional Array that is the destination of the elements copied from WindowCollection.</param>
+        /// <param name="index">The zero-based index in array at which copying begins.</param>
         public void CopyTo(Window[] array, int index)
         {
             _list.CopyTo(array, index);
         }
 
         /// <summary>
-        /// Count property
+        /// Gets the number of windows contained in the WindowCollection.
         /// </summary>
         public int Count
         {
@@ -118,7 +108,7 @@ namespace nanoFramework.UI
         }
 
         /// <summary>
-        /// IsSynchronized
+        /// Gets a value indicating whether access to the WindowCollection is synchronized (thread safe).
         /// </summary>
         public bool IsSynchronized
         {
@@ -129,7 +119,7 @@ namespace nanoFramework.UI
         }
 
         /// <summary>
-        /// SyncRoot
+        /// Gets an object that can be used to synchronize access to the WindowCollection.
         /// </summary>
         public Object SyncRoot
         {
@@ -141,12 +131,11 @@ namespace nanoFramework.UI
 
         #endregion ICollection implementation
 
-        //------------------------------------------------------
-        //
-        //  Internal Methods
-        //
-        //------------------------------------------------------
         #region Internal Methods
+
+        /// <summary>
+        /// Creates a new WindowCollection that is a copy of the current instance.
+        /// </summary>
         internal WindowCollection Clone()
         {
             WindowCollection clone;
@@ -162,16 +151,30 @@ namespace nanoFramework.UI
             return clone;
         }
 
+        /// <summary>
+        /// Removes the specified window from the WindowCollection.
+        /// </summary>
+        /// <param name="win">The window to remove.</param>
         internal void Remove(Window win)
         {
             _list.Remove(win);
         }
 
+        /// <summary>
+        /// Adds the specified window to the WindowCollection.
+        /// </summary>
+        /// <param name="win">The window to add.</param>
+        /// <returns>The index at which the window was added.</returns>
         internal int Add(Window win)
         {
             return _list.Add(win);
         }
 
+        /// <summary>
+        /// Determines whether the specified window is present in the WindowCollection.
+        /// </summary>
+        /// <param name="win">The window to check for.</param>
+        /// <returns>True if the window is present, false otherwise.</returns>
         internal bool HasItem(Window win)
         {
             lock (_list.SyncRoot)
@@ -189,16 +192,5 @@ namespace nanoFramework.UI
         }
 
         #endregion Internal Methods
-
-        //------------------------------------------------------
-        //
-        //  Private Fields
-        //
-        //------------------------------------------------------
-        #region Private Fields
-        private ArrayList _list;
-        #endregion Private Fields
     }
-
-    #endregion WindowCollection class
 }
